+++ packages/deposit-screener/src/cli/commands/run/processor.ts
@@ -20,8 +20,14 @@ const runProcess = new Command("processor")
     "directory to write logs to",
     "./logs/deposit-screener-processor"
   )
+  .option(
+    "--throttle-ms <number>",
+    "maximum period of time to wait before pulling new deposit events",
+    parseInt
+  )
   .action(async (options) => {
-    const { configNameOrPath, logDir } = options;
+    const { configNameOrPath, logDir, throttleMs } = options;
+
     const config = loadNocturneConfig(configNameOrPath);
 
     // TODO: enable switching on adapter impl
@@ -55,10 +61,11 @@ const runProcess = new Command("processor")
       attestationSigner,
       txSigner,
       getRedis(),
-      logger
+      logger,
+      config.contracts.startBlock
     );
 
-    const { promise } = await processor.start();
+    const { promise } = await processor.start(throttleMs);
     await promise;
   });