+++ packages/deposit-screener/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ### Unreleased
 
+- add `--throttle-ms` arg to CLI
+- bump max chunk size up to 100K blocks
+- add optional argument `queryThrottleMs` to `DepositScreenerProcessor.start` and use it when instantiating iterator
+- add support for `throttleMs` option to sync adapter
+- pull `startBlock` from `config.contracts` and pass it to `startBlock` in processor cli
+- add `startBlock` parameter to `DepositScreenerProcessor`
 - subgraph fetch functions query via `idx_gte` and `idx_lt` instead of `id_gte` and `id_lt`
 - create separate `docker-compose.local.yml` and `docker-compose.dev.yml` where `dev` version pulls from docker hub
 - tag docker image with nocturnelabs org name